% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{This is the title.}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\description{
This is the description.
}
\details{
These are further details.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{
add_numbers(1, 2) ## returns 3

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   add_numbers(2, 3)
}

## don't test this during 'R CMD check'
\donttest{
   add_numbers(4, 5)
}
}
