% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackson.R
\docType{data}
\name{jackson}
\alias{jackson}
\title{Jackson Data}
\format{A data frame with 9000 rows and 60 variables:
\describe{
  \item{q0}{Am the life of the party.}
  \item{q1}{Feel little concern for others.}
  \item{q2}{Am always prepared.}
  \item{q3}{Get stressed out easily.}
  \item{q4}{Have a rich vocabulary.}
  \item{q5}{Don't talk a lot.}
  \item{q6}{Am interested in people.}
  \item{q7}{Leave my belongings around.}
  \item{q8}{Am relaxed most of the time.}
  \item{q9}{Have difficulty understanding abstract ideas.}
  \item{q10}{Feel comfortable around people.}
  \item{q11}{Insult people.}
  \item{q12}{Pay attention to details.}
  \item{q13}{Worry about things.}
  \item{q14}{Have a vivid imagination.}
  \item{q15}{Keep in the background.}
  \item{q16}{Sympathize with others' feelings.}
  \item{q17}{Make a mess of things.}
  \item{q18}{Seldom feel blue.}
  \item{q19}{Am not interested in abstract ideas.}
  \item{q20}{Start conversations.}
  \item{q21}{Am not interested in other people's problems.}
  \item{q22}{Get chores done right away.}
  \item{q23}{Am easily disturbed.}
  \item{q24}{Have excellent ideas.}
  \item{q25}{Have little to say.}
  \item{q26}{Have a soft heart.}
  \item{q27}{Often forget to put things back in their proper place.}
  \item{q28}{Get upset easily.}
  \item{q29}{Do not have a good imagination.}
  \item{q30}{Talk to a lot of different people at parties.}
  \item{q31}{Am not really interested in others.}
  \item{q32}{Like order.}
  \item{q33}{Change my mood a lot.}
  \item{q34}{Am quick to understand things.}
  \item{q35}{Don't like to draw attention to myself.}
  \item{q36}{Take time out for others.}
  \item{q37}{Shirk my duties.}
  \item{q38}{Have frequent mood swings.}
  \item{q39}{Use difficult words.}
  \item{q40}{Don't mind being the center of attention.}
  \item{q41}{Feel others' emotions.}
  \item{q42}{Follow a schedule.}
  \item{q43}{Get irritated easily.}
  \item{q44}{Spend time reflecting on things.}
  \item{q45}{Am quiet around strangers.}
  \item{q46}{Make people feel at ease.}
  \item{q47}{Am exacting in my work.}
  \item{q48}{Often feel blue.}
  \item{q49}{Am full of ideas.}
  \item{gender}{Gender}
  \item{age}{Age}
  \item{SecondsElapsed}{SecondsElapsed}
  \item{E}{(30 + q0 - q5 + q10 - q15 + q20 - q25 + q30 - q35 + q40 - q45)/10}
  \item{N}{(24 - q1 + q6 - q11 + q16 - q21 + q26 - q31 + q36 + q41 + q46)/10}
  \item{A}{(24 + q2 - q7 + q12 - q17 + q22 - q27 + q32 - q37 + q42 + q47)/10}
  \item{C}{(12 + q3 - q8 + q13 - q18 + q23 + q28 + q33 + q38 + q43 + q48)/10}
  \item{O}{(18 + q4 - q9 + q14 - q19 + q24 - q29 + q34 + q39 + q44 + q49)/10}
}}
\source{
\url{https://figshare.com/articles/data.csv/96542}
}
\usage{
jackson
}
\description{
Data are already recoded. Do not recode!!!
}
\keyword{datasets}
