% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackson_items.R
\docType{data}
\name{jackson_items}
\alias{jackson_items}
\title{Jackson Data}
\format{A list with the following entries:
\describe{
  \item{N}{N item names}
  \item{E}{E item names}
  \item{O}{O item names}
  \item{A}{A item names}
  \item{C}{C item names}
  \item{rev_item}{reversed item names}
}}
\source{
\url{https://figshare.com/articles/data.csv/96542}
}
\usage{
jackson_items
}
\description{
Data are already recoded. Do not recode!!!
}
\keyword{datasets}
